requirements.txt
asgiref
autopep8
dj-database-url
dj-rest-auth
docker-compose
django
django-cors-headers
django-extra-fields
django-mysql
django-session-timeout
djangorestframework
djangorestframework-jwt
gunicorn
mysqlclient
mysql-connector
python-decouple
pycodestyle
pytz
sqlparse
Unipath 
whitenoise
django-include-by-ajax
django-preventconcurrentlogins
pillow
python-decouple
